Add ignoreNonAnnotated property to TranslatableDTO
It is a boolean value describing behavior regarding fields (or
getters) not annotated with @FieldNameTranslation. Default value is
true', which means that if there is a field (or a getter without
corresponding field) not annotated with @FieldNameTranslation, it will
not be printed by writer. If set to false, then all fields (or
getters) not annotated with @FieldNameTranslation will be printed.
It seems to be a more intuitive behaviour to ignore by default fields
that don't explicitly request translation.